code layout
git-svn-id: http://ltt.polymtl.ca/svn@1350 04897980-b3bd-0310-b5e0-8ef037075253

+++ genevent-new/parser.c
@@ -1228,7 +1228,7 @@ void checkNamedTypesImplemented(table_t * named_types)
  ****************************************************************************/
 
 void generateChecksum(char* facName,
-    unsigned long * checksum, sequence_t * events)
+    unsigned int * checksum, sequence_t * events)
 {
   unsigned long crc ;
   int pos;
